+++ crisp/plugin_one.go
@@ -23,7 +23,7 @@ type PluginInformation struct {
   Type          *string                   `json:"type,omitempty"`
   Name          *string                   `json:"name,omitempty"`
   Description   *string                   `json:"description,omitempty"`
-  Price         *uint                     `json:"price,omitempty"`
+  Price         *uint16                   `json:"price,omitempty"`
   Plans         *[]PluginInformationPlan  `json:"plans,omitempty"`
   Icon          *string                   `json:"icon,omitempty"`
   Configurable  *bool                     `json:"configurable,omitempty"`
@@ -34,7 +34,7 @@ type PluginInformation struct {
 type PluginInformationPlan struct {
   ID     *string  `json:"id,omitempty"`
   Name   *string  `json:"name,omitempty"`
-  Price  *uint    `json:"price,omitempty"`
+  Price  *uint16  `json:"price,omitempty"`
 }
